+++ backend/store/etcd/silenced_store.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
+	"github.com/sensu/sensu-go/backend/etcd"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/backend/store/etcd/kvc"
 	clientv3 "go.etcd.io/etcd/client/v3"
@@ -257,6 +258,7 @@ func (s *Store) arraySilencedEntries(ctx context.Context, resp *clientv3.GetResp
 		if leaseID > 0 {
 			// legacy expiry mechanism
 			ttl, err := s.client.TimeToLive(ctx, leaseID)
+			etcd.LeaseOperationsCounter.WithLabelValues(etcd.LeaseOperationTypeTTL, etcd.LeaseStatusFor(err)).Inc()
 			if err != nil {
 				logger.WithError(err).Error("error setting TTL on silenced")
 				continue
@@ -302,6 +304,7 @@ func (s *Store) arrayTxnSilencedEntries(ctx context.Context, resp *clientv3.TxnR
 			if leaseID > 0 {
 				// legacy expiry mechanism
 				ttl, err := s.client.TimeToLive(ctx, leaseID)
+				etcd.LeaseOperationsCounter.WithLabelValues(etcd.LeaseOperationTypeTTL, etcd.LeaseStatusFor(err)).Inc()
 				if err != nil {
 					logger.WithError(err).Error("error setting TTL on silenced")
 					continue